% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data4500}
\alias{data4500}
\title{Data to use for builiding calibration model (4500, 3 bounce)}
\format{
A data frame with 37 rows, 1773 independent variable and 1 dependent variable
}
\source{
{Created in package predicttransfat to be used for model development}
}
\usage{
data4500
}
\description{
Contains spectra and trans fat values for 37 samples
}
\examples{
data(data4500)

}
\keyword{datasets}
