% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_model.R
\name{transfat_model}
\alias{transfat_model}
\title{Trans fat IR prediction}
\usage{
transfat_model()
}
\value{
calibrated plsr model
}
\description{
This function returns the plsr model obtained using IR data (5500) for trans fat
}
