% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-package.R
\docType{package}
\name{prediction-package}
\alias{margins-package}
\alias{prediction-package}
\title{Tidy, Type-safe \code{prediction()} Methods}
\description{
A one-function package containing \code{\link{prediction}}, a type-safe alernative to \code{\link[stats]{predict}} that always returns a data frame.
}
\author{
Thomas J. Leeper
}
\seealso{
\code{\link{prediction}}
}
\keyword{package}

