% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_model.R
\name{transfat_model}
\alias{transfat_model}
\title{Trans fat IR prediction}
\usage{
transfat_model()
}
\value{
calibrated plsr model
}
\description{
This function returns the calibrated plsr model obtained using IR data (4500 Agilent spectrometer, 3 bounce, 4cm-1 resolution) for trans fat. This model can be used to predict trans fat value of an unknown sample from its IR spectra.
}
\examples{
cal_model <- transfat_model()
}
